Split Italian mobile numbers into 3,4 chunks but allow the subscriber length to range from 6-8 digits.
The standard length of mobile numbers is 10 digits (3 + t subscriber), but as outlined here
http://en.wikipedia.org/wiki/Telephone_numbers_in_Italy#Mobile_telephones, there are 6-digit subscriber numbers
as well as 11. The ITU document for Italy, http://www.itu.int/dms_pub/itu-t/oth/02/02/T020200006B0001PDFE.pdf,
also touches on this on page 14 listing that the minimum number of digits for a mobile number is 9 (here they refer
to ndc + subcriber). Moreover, we also have users with these 9-digit numbers, which is how we actually stumbled
across this problem in the first place.