Extended Estonia number parsing to accept mobile numbers in the 5xx range to have a length of 7 or 8.
The ITU document on Estonia, http://www.itu.int/dms_pub/itu-t/oth/02/02/T02020000430001PDFE.pdf, indicates on
page 4 that the number range 50-59 contains 8 digits, however, if you look at http://en.wikipedia.org/wiki/Telephone_numbers_in_Estonia
and https://www.numberingplans.com/?page=plans&sub=phonenr&alpha_2_input=EE&current_page=69, it appears that this is not
entirely correct. the 5xx range can be either 7 or 8 digits long and this is something that we are also seeing with some of our
clients.

Now to be fair to the ITU document, as an additional remark, it states the following.

"Differences in § 19 section 1 of the “Estonian Numbering Plan” Regulation, Mobile telephone service"

I am not sure what other document this is referring to as the § 19 section 1 is not part of the one pointed to by the aforementioned link.
In either case, given the other sources of information along with some of our clients, it seems that the 7/8 digit is something that
needs to be taken into account.