[Nigeria Mobile Support]: Fixed parsing of Nigerian mobile numbers
References:
[1] Mobile Telephony Number Allocation (Nigerian Communications Commission)
    http://www.ncc.gov.ng/index.php?option=com_content&view=article&id=113&Itemid=102
[2] National Numbering Plan (Nigerian Communications Commission)
    http://www.ncc.gov.ng/index.php?option=com_docman&task=doc_download&gid=92&Itemid=
[3] Telephone numbers in Nigeria (Wiki)
    http://en.wikipedia.org/wiki/Telephone_numbers_in_Nigeria
[4] ITU Nigeria
    http://www.itu.int/dms_pub/itu-t/oth/02/02/T020200009C0001PDFE.pdf

According to the aforementioned references, the area codes for GSM numbers are 3 digits long with the subscriber part having 7 digits. This means the previous match rule
match(/^(702\d)\d+$/)    >> split(3,4) was incorrect as this would allow an 11-digit national number which is not given by any of the specifications. Also, the ranges
700-709, 800-809, 810-819, and 900-909 have been allocated for mobile numbers as outlined in [1].

Fixed up parsing of mobile numbers according to the [1] and extended both the plausibility specs as well as the country specs