[Côte d'Ivoire]: Fixed parsing of mobile numbers and fixed splitting to reflect ITU specification
References:
[0] ITU
    http://www.itu.int/dms_pub/itu-t/oth/02/02/T02020000310001PDFE.pdf
[1] Telephone numbers in Ivory Coast
    http://en.wikipedia.org/wiki/Telephone_numbers_in_Ivory_Coast

Fixes:
The Ivory Coast numbers do not have a trunk code leading to the situation that valid numbers such as '22507123456' would be normalized as '2257123456'. Furthermore, if you look at the
allocation of both landline and mobile numbers given in both [0] and [1], the area code is fixed to 2 digits. In the case of landlines, the ITU spec will list out 3 digit area codes
but this is only because they are making the separation based on telecom provider as well. That is, the 3rd digit of the national subscriber number signifies the network operator. But viewing
it from the zones listed in the "Usage of E.164 number" column, it is clear that the zones are allocated 2-digits. This also matches up with the information in [1] as well as the way the
numbers are formatted in [0]

Côte d'Ivoire
Tel: +225 20 34 43 74
Fax: +225 20 34 43 75
URL: www.artci.ci

As a result, the parsing of Ivory Coast numbers will split the NSN into 4 2-digit numbers. Furthermore, we need to set the trunk code to an empty string to prevent it from being removed
during normalization (similar to how it is done for Italy)