Add "Mass Response" mobile provider to the Austria
This adds the "670" code to the Austria file. "670" is the code for the Mass Response Provider. See [this wikipedia](https://de.wikipedia.org/wiki/Telefonvorwahl_(%C3%96sterreich)#Mobilfunknetze) article (sorry, only in German). Spusu is a provider that launched last year in Austria. [This is their website](https://www.spusu.at/) which shows the code on the frontpage.